Applied isspace(), strip() to user input.
The isspace() function checks if the input string is all whitespaces. If it is the case, then the output will be "Input string is empty".  The strip() function is to remove the whitespaces from the front and back of the input string.

So that  "                     mct                      "  will still be consider a valid input.